+++ app/Http/Controllers/Dashboard/PostController.php
@@ -14,7 +14,12 @@ class PostController extends Controller
 {
 
     use UploadImage;
+    protected $postModel;
 
+    public function __construct(Post $post)
+    {
+        $this->postModel = $post;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -25,24 +30,27 @@ class PostController extends Controller
         return view('dashboard.posts.index');
     }
 
-    public function getPostsDatatable(){
+    public function getPostsDatatable()
+    {
         $posts = Post::select('*')->with('category');
         return Datatables::of($posts)
-        ->addIndexColumn()
-        ->addColumn('title', function ($row) {
-           return $row->translate(app()->getLocale())->title;
-        })
-        ->addColumn('category_name', function ($row) {
-            return $row->category->translate(app()->getLocale())->title;
-        })
-        ->addColumn('action', function ($row) {
-           return $btn = '
-                <a href="' . Route('dashboard.posts.edit', $row->id) . '"  class="edit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a>
-                <a id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm"  data-toggle="modal" data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
-        })
-        
-        ->rawColumns(['title','category_name','action'])
-        ->make(true);
+            ->addIndexColumn()
+            ->addColumn('title', function ($row) {
+                return $row->translate(app()->getLocale())->title;
+            })
+            ->addColumn('category_name', function ($row) {
+                return $row->category->translate(app()->getLocale())->title;
+            })
+            ->addColumn('action', function ($row) {
+                if (auth()->user()->can('update', $row)) {
+
+                    return $btn = '
+                     <a href="' . Route('dashboard.posts.edit', $row->id) . '"  class="edit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a>
+                     <a id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm"  data-toggle="modal" data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
+                }
+            })
+            ->rawColumns(['title', 'category_name', 'action'])
+            ->make(true);
     }
 
     /**
@@ -53,10 +61,10 @@ class PostController extends Controller
     public function create()
     {
         $categories = Category::all();
-        if( count($categories) > 0 ){
-            return view('dashboard.posts.add',compact('categories'));
+        if (count($categories) > 0) {
+            return view('dashboard.posts.add', compact('categories'));
         }
-        abort(404);
+        return redirect()->route('dashboard.categories.create')->with(['error' => 'من فضلك قم بإدخال بعض الأقسام']);
     }
 
     /**
@@ -72,20 +80,19 @@ class PostController extends Controller
         ];
 
         foreach (config('app.languages') as $key => $val) {
-            $data[$key.'*.title'] = 'nullable|string';
-            $data[$key.'*.content'] = 'nullable|string';
-            $data[$key.'*.smallDesc'] = 'nullable|string';
+            $data[$key . '*.title'] = 'nullable|string';
+            $data[$key . '*.content'] = 'nullable|string';
+            $data[$key . '*.smallDesc'] = 'nullable|string';
         }
         $request->validate($data);
 
-        $post = Post::create($request->except('image','_token'));
+        $post = Post::create($request->except('image', '_token'));
 
         if ($request->has('image')) {
             $path = $this->upload($request->image);
             $post->update(['image' => $path]);
         }
         return redirect()->route('dashboard.posts.index');
-
     }
 
     /**
@@ -107,8 +114,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('update',$post);
+
         $categories = Category::all();;
-        return view('dashboard.posts.edit',compact('categories','post'));
+        return view('dashboard.posts.edit', compact('categories', 'post'));
     }
 
     /**
@@ -120,18 +129,20 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $this->authorize('update',$post);
+
         $data = [
             'image' => 'nullable|image|mimes:png,jpg,jpeg,svg,gif|max:2048',
         ];
 
         foreach (config('app.languages') as $key => $val) {
-            $data[$key.'*.title'] = 'nullable|string';
-            $data[$key.'*.content'] = 'nullable|string';
-            $data[$key.'*.smallDesc'] = 'nullable|string';
+            $data[$key . '*.title'] = 'nullable|string';
+            $data[$key . '*.content'] = 'nullable|string';
+            $data[$key . '*.smallDesc'] = 'nullable|string';
         }
         $request->validate($data);
 
-        $post->update($request->except('image','_token'));
+        $post->update($request->except('image', '_token'));
 
         if ($request->has('image')) {
             $path = $this->upload($request->image);
@@ -153,7 +164,9 @@ class PostController extends Controller
 
     public function delete(Request $request)
     {
-        if(is_numeric($request->id)){
+        $this->authorize('delete',$this->postModel->find($request->id));
+
+        if (is_numeric($request->id)) {
             Post::find($request->id)->delete();
         }
         return redirect()->route('dashboard.posts.index');
